Trait again::Condition

pub trait Condition<E> {
    fn is_retryable(&mut self, error: &E) -> bool;
}

A type to determine if a failed Future should be retried

A implementatin is provided for Fn(&Err) -> bool

Required methods

fn is_retryable(&mut self, error: &E) -> bool

Return true if an Futures error is worth retrying

Loading content...