Trait again::Task

pub trait Task {
    type Item;
    type Error: Debug;
    type Fut: Future<Output = Result<Self::Item, Self::Error>>;
    fn call(&mut self) -> Self::Fut;
}

A unit of work to be retried

A implementation is provided for FnMut() -> Future

Associated Types

type Item

The Ok variant of a Futures associated Output type

type Error: Debug

The Err variant of Futures associated Output type

type Fut: Future<Output = Result<Self::Item, Self::Error>>

The resulting Future type

Loading content...
Required methods

fn call(&mut self) -> Self::Fut

Call the operation which invokes results in a Future

Loading content...